+++ packages/datastore/__tests__/AsyncStorageAdapter.test.ts
@@ -13,6 +13,10 @@ let DataStore: typeof DataStoreType;
 const ASAdapter = <any>AsyncStorageAdapter;
 
 describe('AsyncStorageAdapter tests', () => {
+	beforeEach(() => {
+		jest.clearAllMocks();
+	});
+
 	describe('Query', () => {
 		let Model: PersistentModelConstructor<Model>;
 		let model1Id: string;
@@ -61,10 +65,6 @@ describe('AsyncStorageAdapter tests', () => {
 					dateCreated: new Date(baseDate.getTime() + 2).toISOString(),
 				})
 			);
-		});
-
-		beforeEach(() => {
-			jest.clearAllMocks();
 		});
 
 		it('Should call getById for query by id', async () => {
@@ -174,6 +174,52 @@ describe('AsyncStorageAdapter tests', () => {
 			// both should be undefined, even though we only explicitly deleted the user
 			expect(user).toBeUndefined;
 			expect(profile).toBeUndefined;
+		});
+	});
+
+	describe('Save', () => {
+		let User: PersistentModelConstructor<User>;
+		let Profile: PersistentModelConstructor<Profile>;
+		let profile: Profile;
+
+		beforeAll(async () => {
+			({ initSchema, DataStore } = require('../src/datastore/datastore'));
+
+			const classes = initSchema(testSchema());
+
+			({ User } = classes as {
+				User: PersistentModelConstructor<User>;
+			});
+
+			({ Profile } = classes as {
+				Profile: PersistentModelConstructor<Profile>;
+			});
+
+			profile = await DataStore.save(
+				new Profile({ firstName: 'Rick', lastName: 'Bob' })
+			);
+		});
+
+		it('should allow linking model via model field', async () => {
+			const savedUser = await DataStore.save(
+				new User({ name: 'test', profile })
+			);
+			const user1Id = savedUser.id;
+
+			const user = await DataStore.query(User, user1Id);
+			expect(user.profileID).toEqual(profile.id);
+			expect(user.profile).toEqual(profile);
+		});
+
+		it('should allow linking model via FK', async () => {
+			const savedUser = await DataStore.save(
+				new User({ name: 'test', profileID: profile.id })
+			);
+			const user1Id = savedUser.id;
+
+			const user = await DataStore.query(User, user1Id);
+			expect(user.profileID).toEqual(profile.id);
+			expect(user.profile).toEqual(profile);
 		});
 	});
 });