+++ packages/datastore/src/storage/adapter/IndexedDBAdapter.ts
@@ -119,8 +119,10 @@ class IndexedDBAdapter implements Adapter {
 										autoIncrement: true,
 									});
 
-									const indexes = this.schema.namespaces[namespaceName]
-										.relationships[modelName].indexes;
+									const indexes =
+										this.schema.namespaces[namespaceName].relationships[
+											modelName
+										].indexes;
 									indexes.forEach(index => store.createIndex(index, index));
 
 									store.createIndex('byId', 'id', { unique: true });
@@ -547,9 +549,9 @@ class IndexedDBAdapter implements Adapter {
 			const storeName = this.getStorenameForModel(modelConstructor);
 
 			const models = await this.query(modelConstructor, condition);
-			const relations = this.schema.namespaces[nameSpace].relationships[
-				modelConstructor.name
-			].relationTypes;
+			const relations =
+				this.schema.namespaces[nameSpace].relationships[modelConstructor.name]
+					.relationTypes;
 
 			if (condition !== undefined) {
 				await this.deleteTraverse(
@@ -625,9 +627,9 @@ class IndexedDBAdapter implements Adapter {
 				}
 				await tx.done;
 
-				const relations = this.schema.namespaces[nameSpace].relationships[
-					modelConstructor.name
-				].relationTypes;
+				const relations =
+					this.schema.namespaces[nameSpace].relationships[modelConstructor.name]
+						.relationTypes;
 
 				await this.deleteTraverse(
 					relations,
@@ -637,9 +639,9 @@ class IndexedDBAdapter implements Adapter {
 					deleteQueue
 				);
 			} else {
-				const relations = this.schema.namespaces[nameSpace].relationships[
-					modelConstructor.name
-				].relationTypes;
+				const relations =
+					this.schema.namespaces[nameSpace].relationships[modelConstructor.name]
+						.relationTypes;
 
 				await this.deleteTraverse(
 					relations,
@@ -723,12 +725,15 @@ class IndexedDBAdapter implements Adapter {
 
 						const hasOneCustomField = targetName in model;
 						const value = hasOneCustomField ? model[targetName] : model.id;
+						if (!value) break;
 
-						const recordToDelete = <T>await this.db
-							.transaction(storeName, 'readwrite')
-							.objectStore(storeName)
-							.index(hasOneIndex)
-							.get(value);
+						const recordToDelete = <T>(
+							await this.db
+								.transaction(storeName, 'readwrite')
+								.objectStore(storeName)
+								.index(hasOneIndex)
+								.get(value)
+						);
 
 						await this.deleteTraverse(
 							this.schema.namespaces[nameSpace].relationships[modelName]