+++ packages/api-graphql/CHANGELOG.md
@@ -3,24 +3,49 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [2.2.12](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.2.11...@aws-amplify/api-graphql@2.2.12) (2021-11-09)
+## [2.2.20](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.2.19...@aws-amplify/api-graphql@2.2.20) (2022-01-27)
 
+**Note:** Version bump only for package @aws-amplify/api-graphql
+
+## [2.2.19](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.2.18...@aws-amplify/api-graphql@2.2.19) (2022-01-07)
+
+**Note:** Version bump only for package @aws-amplify/api-graphql
+
+## [2.2.18](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.2.17...@aws-amplify/api-graphql@2.2.18) (2021-12-16)
+
+### Bug Fixes
+
+- **@aws-amplify/api-graphql:** Fix webpack build ([#9358](https://github.com/aws-amplify/amplify-js/issues/9358)) ([73587d7](https://github.com/aws-amplify/amplify-js/commit/73587d78fc4feba39d5a527bfe2c790935653fd3))
+
+## [2.2.17](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.2.16...@aws-amplify/api-graphql@2.2.17) (2021-12-03)
+
+**Note:** Version bump only for package @aws-amplify/api-graphql
+
+## [2.2.16](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.2.15...@aws-amplify/api-graphql@2.2.16) (2021-12-02)
+
+**Note:** Version bump only for package @aws-amplify/api-graphql
+
+## [2.2.15](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.2.14...@aws-amplify/api-graphql@2.2.15) (2021-11-18)
+
+**Note:** Version bump only for package @aws-amplify/api-graphql
+
+## [2.2.14](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.2.13...@aws-amplify/api-graphql@2.2.14) (2021-11-16)
+
+**Note:** Version bump only for package @aws-amplify/api-graphql
+
+## [2.2.13](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.2.12...@aws-amplify/api-graphql@2.2.13) (2021-11-12)
+
+**Note:** Version bump only for package @aws-amplify/api-graphql
+
+## [2.2.12](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.2.11...@aws-amplify/api-graphql@2.2.12) (2021-11-09)
 
 ### Reverts
 
-* uuid dependency upgrade ([#9159](https://github.com/aws-amplify/amplify-js/issues/9159)) ([4ef8aa9](https://github.com/aws-amplify/amplify-js/commit/4ef8aa9c7c25dbe921fd02b6205b8defb93fbaec))
-
-
-
-
+- uuid dependency upgrade ([#9159](https://github.com/aws-amplify/amplify-js/issues/9159)) ([4ef8aa9](https://github.com/aws-amplify/amplify-js/commit/4ef8aa9c7c25dbe921fd02b6205b8defb93fbaec))
 
 ## [2.2.11](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.2.10...@aws-amplify/api-graphql@2.2.11) (2021-10-28)
 
 **Note:** Version bump only for package @aws-amplify/api-graphql
-
-
-
-
 
 ## [2.2.10](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/api-graphql@2.2.9...@aws-amplify/api-graphql@2.2.10) (2021-10-21)